+++ utils/src/output_bytes.rs
@@ -2,7 +2,7 @@
 ///
 /// # Arguments
 /// * `v` - the size in bytes
-pub fn output_bytes(v: usize) -> String {
+pub fn output_bytes(v: u64) -> String {
     let map = vec![
         (1_099_511_627_776, "TiB"),
         (1_073_741_824, "GiB"),
